refactor(core): make TurnOrigin single-authority
Move the shared TurnOrigin type and decoder into a leaf module and use it from stored-message and RuntimeEvent decoding.

Generated-by: Codex
